ecore: Add a new API function ecore_main_loop_nested_get
Summary: This commit adds a new function 'ecore_main_loop_nested_get'
so that we can detect if the ecore_main_loop is running.

NB: This is going to be needed for a new eldbus function that we have
to add in order to handle a use-case on the Wayland side. Spoke with
cedric for a while wrt to all this, and he gave it his 'ok' ;)

@feature

Signed-off-by: Chris Michael <cp.michael@samsung.com>