DnD: fix callbacks deletion for inline windows.
During deletion of a window, widgets considered as droppable targets
have to remove their DnD callbacks. To achieve this, elm_drop_target_del
is called from the DEL callback (destructor). This function has to
determine if X11 or Wayland is used. Since the parent is already unknown
at this stage, only checking the engine name can give this
information.
On a regular window, the engine name is related to the target display.

The problem happens when an inline window is used. The engine is a
buffer and no information is given regarding the target display.

The patch fixes it by checking the nature of the Ecore Evas parents. It
supports nested windows (inline inside inline... inside XWin).

@fix