tests/ecore_thread: Disable verbose logging with printf
If we are going to test for 10000000 messages we should not print out for each
of them to stdout. The log file for ecore_suite will catch them all. It ended
up being 199MB on my machine just filled with these debug statements.

I left it in in case someone needs it for debugging but we should not have it
on by default.