elm - elm image orientation api fix - major bug in implementation
@fix

The elm image orient_set api looks like it is a state, and it is
documented to SET the orietnation, but the code actually DOEs that
orientation when you set it. so docs and api name didnt match
functionality at all. this was bad an an oversight. you didn't notice
until you set orientation multiple times to an image. this fixes the
implementation to be as documented - a state that is applied and
remains that way until set to something else. setting to the same
orientation has no efect as with color, size, etc. etc. in evas.

you could look on this as an api break, but then either the docs
change AND api name changes to match the code, or code changes to
match the docs. we don't keep bug compatibility in efl and in this
case i'm making the call to break the behavior of a  function as it
was broken already with respect to name and docs.