Edje: edje_cc - abort recursive Reference that is made by GROUP parts
Summary:
It is easy to create edj collections that aren't working at all and
edje_cc easily allows user to do that.
For example:
>   Having group A with GROUP part that has group B as source.
>   Having group B with GROUP part that has group A as source.
In this case edje_cc compile source code perfectly, but if user try to load
this edje as layout or use together with edje_edit it will cause unexpacted
and wrong behaviour.

@fix

Reviewers: seoz, Hermet, reutskiy.v.v, cedric

Reviewed By: cedric

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D1908

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>