ecore_win32: don't allocate Ecore_Event_Key events with NULL keys
If "keyname" and "key" are NULL an Ecore_Event_Key is allocated
throught calloc. strlen() is called on these variables, which crashes
the program. As it is not correct to return a key event for NULL
keys to ecore, it is preferrable to return NULL and exit the function
before the key event is allocated, which fixes the crash in the same
time.

@fix

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>