edje: epp - fix build of edje file in cas of recursive includes.
Amitesh Singh fixed this issue with Commit
e66e132d9dabecb8273fe3962099d4523fef1055, but this commit has been reverted
recently, by a commit freeing ressources
(6d0b834a76c3301bf4c4b29752f902dbe8a8525e). It's clear that we are leaking here
if we remove the free, but there is also a valgrind error when this memory is
freed. After trying to debug and understand the code of epp, i can only
propose to remove this free, we will need to include a lot of files before this
memory leak could be visible.

It is also not a real problem, the memory leak only concern the file name passed
in the include directive and epp is started per file anyway. It is unlikely to
create any real life problem and we better spend our time on more pressing matter.

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>