DnD/X11: improve types checking during position event.
During a DnD position event, we need to find the first dropable target
located under the mouse pointer; one of its callbacks has to support
at least one of the types of the dragged data.
This search process was done by strcmp-ing types for every drop target
and for each of their types until a good target is found.
This patch improves the process by comparing types formats, i.e and-ing bits
vectors.