build: Workaround an automake limitation during parallel install relinking
With this commit I'm finally able to use -j10 for make install on my machine.

During install libtool does some relinking which can result in to broken linking
if the dependencies are not handled correctly. Sadly automake has a problem with
the automatic dependency handling during install with LTLIBRARIES which we use
for all our modules. For the details please see this 4.5 years old bug report:
http://debbugs.gnu.org/cgi/bugreport.cgi?bug=7328

We are now setting the dependency manually to force automake to the right decision
during install relinking.

Speed improvement itself is not that high (make -j 1 compared to -j10):
real 0m21.410s  vs. real 0m17.066s

The bigger benefit is the unified use of MAKEOPTS or normal -j X in all our
build targets. I have seen quite some bug reports where -j was used for install
target when it was used in the build target. Last but not least it helps me to
unify some parts of the jenkins jobs and finally allows me to run distcheck
with -j Which uses install internally and failed before. Which goes down from
real 12m50.349s to real 5m52.120s.