Edje: edje edit - ability to remove last item in BOX/TABLE
Removing last item in BOX/TABLE part is actually successful
but then we shouldn't realloc an array of items for 0 items.
That's ridiculous
(and because of that function removes EINA_FALSE,
so user could think that it's unable to remove last item).

So simply array can be set into NULL.

@fix