Gengrid: Modified to handle the focus highlight on unrealized items
This commit is related to 260cdd6150d567fcee8d1afaf690e4b90cd16a2b.

When unrealized item is focused, edje object is not exist,
so genlist cannot read the focus highlight information from edje object.
Therefore, when the item is realized, check the item is focused or not and
need to update focus highlight.